+++ ai4rag/components/data/__init__.py
@@ -4,13 +4,14 @@
 # -----------------------------------------------------------------------------
 from ai4rag.components.data.documents_discovery import DiscoveryResult, DocumentDescriptor, discover_documents
 from ai4rag.components.data.test_data_loader import TestDataLoaderError, TestDataResult, load_test_data
-from ai4rag.components.data.text_extraction import ExtractionResult, extract_text
+from ai4rag.components.data.text_extraction import DoclingExtractionConfig, ExtractionResult, extract_text
 
 __all__ = [
     "discover_documents",
     "DiscoveryResult",
     "DocumentDescriptor",
     "extract_text",
+    "DoclingExtractionConfig",
     "ExtractionResult",
     "load_test_data",
     "TestDataLoaderError",